requirements.txt
Flask>=2.2.5
Flask-Cors>=3.1.0
mysql-connector-python>=8.0.33
bcrypt>=4.0.1
gunicorn>=20.1.0
